+++ src/main/scala/com/guizmaii/scalajwt/implementations/ConfigurableJwtValidator.scala
@@ -1,23 +1,27 @@
 package com.guizmaii.scalajwt.implementations
 
-import java.text.ParseException
-
 import com.guizmaii.scalajwt._
-import com.guizmaii.scalajwt.utils.SupportedJWSAlgorithms._
 import com.nimbusds.jose.jwk.source.JWKSource
 import com.nimbusds.jose.proc.{JWSVerificationKeySelector, SecurityContext}
 import com.nimbusds.jwt.JWTClaimsSet
-import com.nimbusds.jwt.proc.{BadJWTException, DefaultJWTClaimsVerifier, DefaultJWTProcessor}
+import com.nimbusds.jwt.proc.{BadJWTException, DefaultJWTProcessor, JWTClaimsSetVerifier}
 
-import scala.collection.compat.immutable._
+import java.text.ParseException
+import scala.util.control.NonFatal
 
 object ConfigurableJwtValidator {
   def apply(
       keySource: JWKSource[SecurityContext],
+      claimsVerifier: JWTClaimsSetVerifier[SecurityContext],
       algorithm: SupportedJWSAlgorithm = RS256,
-      maybeCtx: Option[SecurityContext] = None,
-      additionalValidations: List[(JWTClaimsSet, SecurityContext) => Option[BadJWTException]] = List.empty
-  ): ConfigurableJwtValidator = new ConfigurableJwtValidator(keySource, algorithm, maybeCtx, additionalValidations)
+      maybeCtx: Option[SecurityContext] = None
+  ): ConfigurableJwtValidator =
+    new ConfigurableJwtValidator(
+      keySource = keySource,
+      claimsVerifier = claimsVerifier,
+      algorithm = algorithm,
+      maybeCtx = maybeCtx
+    )
 }
 
 /** A (fully?) configurable JwtValidator implementation.
@@ -26,14 +30,15 @@ object ConfigurableJwtValidator {
   *   https://connect2id.com/products/nimbus-jose-jwt/examples/validating-jwt-access-tokens
   *
   * @param keySource (Required) JSON Web Key (JWK) source.
+  * @param claimsVerifier (Required) The claims validation rules. See https://connect2id.com/products/nimbus-jose-jwt/examples/validating-jwt-access-tokens#claims
+  * @param algorithm (Optional) Algorithm used to encrypt the token. Default is `RS256`.
   * @param maybeCtx (Optional) Security context. Default is `null` (no Security Context).
-  * @param additionalValidations (Optional) List of additional validations that will be executed on the JWT token. Default is an empty List.
   */
-final class ConfigurableJwtValidator(
+final class ConfigurableJwtValidator private[scalajwt] (
     keySource: JWKSource[SecurityContext],
+    claimsVerifier: JWTClaimsSetVerifier[SecurityContext],
     algorithm: SupportedJWSAlgorithm = RS256,
-    maybeCtx: Option[SecurityContext] = None,
-    additionalValidations: List[(JWTClaimsSet, SecurityContext) => Option[BadJWTException]] = List.empty
+    maybeCtx: Option[SecurityContext] = None
 ) extends JwtValidator {
 
   // Set up a JWT processor to parse the tokens and then check their signature
@@ -43,36 +48,19 @@ final class ConfigurableJwtValidator(
   // RSA keys sourced from the JWK set URL
   private val keySelector = new JWSVerificationKeySelector[SecurityContext](algorithm.nimbusRepresentation, keySource)
   jwtProcessor.setJWSKeySelector(keySelector)
-
-  // Set the additional validations.
-  //
-  // Updated and adapted version of this example:
-  //   https://connect2id.com/products/nimbus-jose-jwt/examples/validating-jwt-access-tokens#claims-validator
-  jwtProcessor.setJWTClaimsSetVerifier(new DefaultJWTClaimsVerifier[SecurityContext] {
-    override def verify(claimsSet: JWTClaimsSet, context: SecurityContext): Unit = {
-      super.verify(claimsSet, context)
-
-      additionalValidations
-        .to(LazyList)
-        .map(f => f(claimsSet, context))
-        .collect { case Some(e) => e }
-        .foreach(e => throw e)
-    }
-  })
+  jwtProcessor.setJWTClaimsSetVerifier(claimsVerifier)
 
   private val ctx: SecurityContext = maybeCtx.orNull
 
-  override def validate(jwtToken: JwtToken): Either[BadJWTException, (JwtToken, JWTClaimsSet)] = {
+  override def validate(jwtToken: JwtToken): Either[InvalidToken, JWTClaimsSet] = {
     val content: String = jwtToken.content
-    if (content.isEmpty) Left(EmptyJwtTokenContent)
+    if (content.isEmpty) Left(InvalidToken(new RuntimeException("Empty JWT token")))
     else
-      try {
-        val claimsSet = jwtProcessor.process(content, ctx)
-        Right(jwtToken -> claimsSet)
-      } catch {
-        case e: BadJWTException => Left(e)
-        case _: ParseException  => Left(InvalidJwtToken)
-        case e: Exception       => Left(UnknownException(e))
+      try Right(jwtProcessor.process(content, ctx))
+      catch {
+        case e: BadJWTException => Left(InvalidToken(e))
+        case e: ParseException  => Left(InvalidToken(e))
+        case NonFatal(e)        => Left(InvalidToken(e))
       }
   }
 }